Feature: refine subscription management UI/UX (#7557)
