+++ src/DonorDashboards/resources/js/app/components/button/style.scss
@@ -6,9 +6,10 @@
     padding: 12px 20px;
     font-size: 16px;
     font-weight: 600;
-    border: 1px solid #0073aa;
+    border: 1px solid var(--give-donor-dashboard-accent-color);
     border-radius: 3px;
     display: inline-flex;
+    width: fit-content;
     align-items: center;
     box-shadow: 0 0 0 0 #7ec980, 0 0 0 0 #4fa651;
     transition: box-shadow 0.1s ease, background-color ease-in 0.3s;
@@ -28,28 +29,53 @@
     }
 
     &.give-donor-dashboard-button--primary {
-        flex: 1;
-        max-width: 240px;
-        border: 1px solid #0073aa;
         color: #fff !important;
-        background: #0073aa;
+        position: relative;
+        background: none;
         box-shadow: none;
         justify-content: center;
+        overflow: hidden;
+
+        &:before {
+            content: '';
+            position: absolute;
+            top: 0;
+            right: 0;
+            bottom: 0;
+            left: 0;
+            background: var(--give-donor-dashboard-accent-color);
+            z-index: 0;
+            transition: filter 0.3s ease;
+        }
+
+        &:hover:before {
+            filter: brightness(90%); // Darkens the background on hover
+        }
+
+        &.disabled:before {
+            display: none;
+        }
+
+        span {
+            position: relative;
+            z-index: 1;
+        }
     }
 
     &.give-donor-dashboard-button--variant {
-        flex: 1;
-        width: 240px;
-        color: #0073aa !important;
+        color: var(--give-donor-dashboard-accent-color) !important;
         background: var(--givewp-shades-white);
         box-shadow: none;
-        border: 1px solid #0073aa;
+        border: 1px solid var(--give-donor-dashboard-accent-color);
         margin: 0;
         justify-content: center;
 
         &:hover {
-            background: #0073aa;
-            color: #fff !important;
+            filter: brightness(90%);
+        }
+
+        span {
+            color: inherit;
         }
     }
 }
