+++ src/DonorDashboards/resources/js/app/components/subscription-manager/amount-control/index.js
@@ -99,14 +99,12 @@ const AmountControl = ({currency, onChange, value, options, min, max}) => {
     return (
         <div className="give-donor-dashboard-amount-inputs">
             <FieldRow>
-                <div>
-                    <SelectControl
+                <SelectControl
                         label={__('Subscription Amount', 'give')}
                         options={options}
                         value={selectValue}
                         onChange={setSelectValue}
                     />
-                </div>
                 <div>
                     {selectValue === CUSTOM_AMOUNT && (
                         <div className="give-donor-dashboard-currency-control">
