--- src/DonorDashboards/resources/js/app/components/subscription-manager/field-content/index.tsx
@@ -1,5 +1,0 @@
-import './style.scss';
-
-export function FieldContent({classNames, children}) {
-    return <div className={`give-donor-dashboard-field-content ${classNames}`}>{children}</div>;
-}
