--- src/DonorDashboards/resources/js/app/components/subscription-manager/field-content/style.scss
@@ -1,7 +1,0 @@
-.give-donor-dashboard-field-content {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 22.5rem;
-    padding-bottom: 5rem;
-}
