+++ src/DonorDashboards/resources/js/app/components/subscription-manager/subscription-status/index.tsx
@@ -5,16 +5,12 @@ import {__} from '@wordpress/i18n';
 import './style.scss';
 
 export default function SubscriptionStatus({subscription}) {
-    const isActive = subscription.payment.status.id === 'active';
+    const status = subscription.payment.status.id;
+    const label = subscription.payment.status.label;
 
     return (
-        <div
-            className={cx('givewp-dashboard-subscription-status', {
-                ['givewp-dashboard-subscription-status--active']: isActive,
-                ['givewp-dashboard-subscription-status--paused']: !isActive,
-            })}
-        >
-            {isActive ? __('Active', 'give') : __('Paused', 'give')}
+        <div className={`givewp-dashboard-subscription-status givewp-dashboard-subscription-status--${status}`}>
+            {label}
         </div>
     );
 }
