+++ src/DonorDashboards/resources/js/app/components/subscription-row/style.scss
@@ -1,0 +1,5 @@
+#give-donor-dashboard {
+    .give-donor-dashboard-table__donation-receipt__cancel {
+        color: #d92d0b;
+    }
+}
